chore: cleaned up errors folder structure and placed elements within standard practice
