+++ src/App.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { AuthProvider } from "./context/AuthContext";
-import Login from "./components/Auth/Login";
-import Dashboard from "./components/pages/Dashboard";
+import Login from "./Auth/Login";
+import Dashboard from "./pages/Dashboard";
 import PrivateRoute from "./components/PrivateRoute";
-import Home from "./components/Home";
+import Home from "./pages/Home";
 import Navigation from "./components/Navigation";
-import Signup from "./components/Auth/Signup";
+import Signup from "./Auth/Signup";
 import Footer from "./components/Footer";
 
 const App: React.FC = () => {
