--- src/components/Home.tsx
@@ -1,19 +1,0 @@
-import CalculateSection from "./CalculateSection"
-import Hero from "./Hero"
-import SearchSection from "./SearchSection"
-import Testimonials from "./Testimonials"
-
-
-const Home = () => {
-  return (
-    <>
-    <Hero />
-    <CalculateSection />
-    <SearchSection />
-    <Testimonials />
-
-    </>
-  )
-}
-
-export default Home
