+++ src/context/AuthContext.tsx
@@ -1,6 +1,6 @@
 // src/context/AuthContext.tsx
 import React, { createContext, useContext, useState, ReactNode } from "react";
-import { auth } from "../firebase/firebaseConfig";
+import { auth } from "../lib/firebase/firebaseConfig";
 import { signOut, onAuthStateChanged, User } from "firebase/auth";
 
 interface AuthContextType {
