+++ src/pages/Home.tsx
@@ -1,0 +1,19 @@
+import CalculateSection from "../components/CalculateSection"
+import Hero from "../components/Hero"
+import SearchSection from "../components/SearchSection"
+import Testimonials from "../components/Testimonials"
+
+
+const Home = () => {
+  return (
+    <>
+    <Hero />
+    <CalculateSection />
+    <SearchSection />
+    <Testimonials />
+
+    </>
+  )
+}
+
+export default Home
